Started branch for C3 dev

+++ Firmware/smog-alado/src/shared/variables.h
@@ -41,7 +41,7 @@ uint16_t idleMinutes = 0;
 uint16_t idleSeconds = 0;
 uint16_t remainingTime = 100;
 uint16_t APtimeout = 180;
-uint16_t minutes = 10;
+uint32_t minutes = 10;
 char timeStr[20];
 
 //* Booleans to check if functions will be called
